Fixes CAA record in Terraform
TL;DR
-----

Updates the data in the CAA record to prevent spurious mismatches

Details
-------

The value of the CAA record was applying correctly but GCP was
making an update to put the `letsencrypt.org` in quotes. This led
to spurious "changes" on plan/apply. Updates the value to only
show real changes.